Add pagination
Since I want to reduce the number of object allocations I don't want to
map all the items from the response into Ruby Objects (PORO)

Therefore the gems that expect pagination assume that we are working
with ActiveRecord objects.

Since I have a list of hashes I need something agnostic of my data.
Pagy worked very well to just render the appropiate component if the
configuration of total and actual page is passed.